Add inflation page and correct quantile labels on charts (#103)
This pull request adds a new page for displaying inflation data and
corrects the quantile labels on the charts. The inflation page provides
information on the Consumer Price Index used by the Reserve Bank of
Australia to index inflation. It includes visualizations of long-term
trends and percentile YoY changes. The quantile labels on the charts
have been adjusted to accurately reflect the data. The changes also
include a link to the Jupyter Notebook used for generating the inflation
data.