Update index.md and retirement.ipynb, and refactor chart_format.py (#168)
This pull request includes updates to the index.md and retirement.ipynb
files, as well as a refactoring of the chart_format.py file. The changes
in index.md involve updating the link to the Tesla (TSLA) page, and the
retirement.ipynb file now includes a reference to a personalized
retirement planner calculator published by the Australian government.
The refactoring of chart_format.py includes the creation of two new
classes, BaseChart and LogChart, which provide additional functionality
for creating charts with logarithmic y-axis scales.