Update schedule for running notebooks (#173)
This pull request updates the schedule for running notebooks. The previous schedule was set to run every Thursday after extended hours, but it has been changed to run every day at 1:00 AM UTC. This change ensures that the notebooks are refreshed regularly and the cache data is up to date.