Add human_format function to LogChart class (#199)
This pull request adds a new function called `human_format` to the `LogChart` class. This function is used to format the y-axis tick labels in a logarithmic chart. It converts large numbers into a more readable format, such as "1k" for 1000, "1m" for 1 million, etc. This improves the readability of the chart for users.